pyproject.toml
[project]
name = "jarvis-va"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "pyaudio>=0.2.14",
    "pyinstaller>=6.14.1",
    "python-dotenv>=1.1.1",
    "pyttsx3>=2.98",
    "speechrecognition>=3.14.3",
]
